展示用户信息出现问题，需要调试

+++ src/main/java/com/baojie/admin/jpa/executor/UserExec.java
@@ -110,7 +110,10 @@ public class UserExec {
         if (ids.length <= 0) {
             return;
         }
-        for (long id : ids) {
+        for (Long id : ids) {
+            if (null == id) {
+                continue;
+            }
             try {
                 repo.deleteById(id);
             } catch (Throwable te) {
